+++ src/main/java/com/example/model/event/EventUpdate.java
@@ -11,5 +11,5 @@ public class EventUpdate {
     private String name;
     private String location;
     private LocalDateTime date;
-    private List<Long> attendees;
+    private List<Integer> attendees;
 }
