+++ src/main/java/com/example/model/friendship/FriendshipRequest.java
@@ -4,7 +4,7 @@ import lombok.*;
 
 @Data
 public class FriendshipRequest {
-    private Long user1Id;
-    private Long user2Id;
+    private Integer user1Id;
+    private Integer user2Id;
     private String status;
 }
