+++ src/main/java/com/example/model/group/GroupRequest.java
@@ -6,7 +6,7 @@ import java.util.List;
 @Data
 public class GroupRequest {
     private String name;
-    private List<Long> members;
+    private List<Integer> members;
 }
