+++ src/main/java/com/example/repository/ConnectionRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 
 
-public interface ConnectionRepository extends JpaRepository<Connection, Long> {
+public interface ConnectionRepository extends JpaRepository<Connection, Integer> {
 }
