+++ src/main/java/com/example/repository/FriendshipRepository.java
@@ -4,6 +4,6 @@ import com.example.entity.Friendship;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface FriendshipRepository extends JpaRepository<Friendship, Long> {
+public interface FriendshipRepository extends JpaRepository<Friendship, Integer> {
 }
