+++ src/main/java/com/example/repository/GroupRepository.java
@@ -4,5 +4,5 @@ import com.example.entity.Group;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface GroupRepository extends JpaRepository<Group, Long> {
+public interface GroupRepository extends JpaRepository<Group, Integer> {
 }
